Revert "Refactor hero control system"
