+++ Assets/Prefabs/Warrior.prefab
@@ -337,9 +337,6 @@ GameObject:
   - component: {fileID: 8560947185031467494}
   - component: {fileID: 3963385185172838639}
   - component: {fileID: 8421911332518481540}
-  - component: {fileID: 3947118601378384970}
-  - component: {fileID: 7405545789030251085}
-  - component: {fileID: 2723759396440014334}
   m_Layer: 7
   m_Name: Warrior
   m_TagString: Hero
@@ -519,42 +516,6 @@ MonoBehaviour:
   wallAvoidFalloff: 1
   priority: 0.223
   debug: 0
---- !u!114 &3947118601378384970
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1941723749653959437}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 10fc6793d0ce4ef0b8c208d4fa52d504, type: 3}
-  m_Name:
-  m_EditorClassIdentifier:
---- !u!114 &7405545789030251085
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1941723749653959437}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b1144c4c8b324fda8b4bf66489e6b39b, type: 3}
-  m_Name:
-  m_EditorClassIdentifier:
---- !u!114 &2723759396440014334
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1941723749653959437}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2f536a7bcb124d02ace1ba2b21b52562, type: 3}
-  m_Name:
-  m_EditorClassIdentifier:
 --- !u!1 &6664528133109611933
 GameObject:
   m_ObjectHideFlags: 0
