--- Assets/Scripts/Hero/CombatController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b1144c4c8b324fda8b4bf66489e6b39b
