--- Assets/Scripts/Hero/HeroStateMachine.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2f536a7bcb124d02ace1ba2b21b52562
