--- Assets/Scripts/Hero/MovementController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 10fc6793d0ce4ef0b8c208d4fa52d504
